+++ .prettierignore
@@ -14,3 +14,4 @@ storybook-static
 .firebase
 functions
 .vercel
+gql
