+++ package.json
@@ -13,21 +13,22 @@
     "format": "prettier --write .",
     "test-all": "yarn check-lint && yarn check-types",
     "test": "jest",
-    "codegen": "graphql-codegen --config codegen.js",
+    "codegen": "graphql-codegen --config codegen.ts",
     "analyze": "ANALYZE=true next build",
     "sb": "storybook dev -p 6006",
     "sb:build": "storybook build",
     "test:sb": "test-storybook"
   },
   "dependencies": {
+    "@graphql-codegen/near-operation-file-preset": "^2.5.0",
     "@headlessui/react": "^1.7.14",
     "@headlessui/tailwindcss": "^0.1.2",
     "@heroicons/react": "^2.0.13",
     "@hookform/resolvers": "^3.1.0",
     "@next/bundle-analyzer": "^13.1.6",
     "@next/font": "13.1.1",
-    "@tanstack/react-query": "^4.20.4",
-    "@tanstack/react-query-devtools": "^4.20.4",
+    "@tanstack/react-query": "^5.0.0-beta.7",
+    "@tanstack/react-query-devtools": "^5.0.0-beta.9",
     "@types/node": "18.11.18",
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
@@ -46,6 +47,7 @@
     "react-dom": "18.2.0",
     "react-error-boundary": "^4.0.10",
     "react-hook-form": "^7.43.9",
+    "react-intersection-observer": "^9.5.2",
     "react-textarea-autosize": "^8.4.0",
     "tailwind-merge": "^1.10.0",
     "typescript": "4.9.4",
@@ -54,12 +56,9 @@
   },
   "devDependencies": {
     "@cloudflare/next-on-pages": "^1.2.0",
-    "@graphql-codegen/cli": "4.0.0",
-    "@graphql-codegen/client-preset": "4.0.0",
-    "@graphql-codegen/near-operation-file-preset": "^2.5.0",
-    "@graphql-codegen/typescript": "^4.0.0",
-    "@graphql-codegen/typescript-operations": "^4.0.0",
-    "@graphql-codegen/typescript-react-query": "^4.1.0",
+    "@graphql-codegen/cli": "5.0.0",
+    "@graphql-codegen/client-preset": "4.1.0",
+    "@graphql-codegen/introspection": "4.0.0",
     "@storybook/addon-a11y": "^7.0.9",
     "@storybook/addon-essentials": "^7.0.9",
     "@storybook/addon-interactions": "^7.0.9",
@@ -71,6 +70,7 @@
     "@storybook/testing-library": "^0.0.14-next.2",
     "@svgr/webpack": "^6.5.1",
     "@tailwindcss/forms": "^0.5.3",
+    "@tanstack/eslint-plugin-query": "^5.0.0-beta.10",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
