+++ tsconfig.json
@@ -20,7 +20,7 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "codegen.js",
+    "codegen.ts",
     "jest.config.ts"
   ],
   "exclude": ["node_modules"]
